Add control for online and offline cpus
Add a toggle button to enable or disable a cpu.
If a cpu is offline, the frequency sliders are disabled.

Some cpus cannot be disabled. In this case the checkbox
is disabled not to confuse the user.

Add a refresh button next to `Apply` to reload the settings
from disk.