Fix dbus error when calling _is_authorized
If the helper was running but the polkit daemon had gone through a
restart the `_is_authorized` would return a `dbus.exceptions.DBusException`.
The error was `org.freedesktop.DBus.Error.ServiceUnknown`.

The issue was that the proxy object for the authority, as well as the
authority itself, were instantiated once at `init_polkit`.

To fix the issue the proxy and the authority were moved inside the
`_is_authorized` method. Thus, everytime the function is called it will
get a fresh reference.